распределение компонентов по структуре FSD
