--- src/components/Card/Card.jsx
@@ -1,30 +1,0 @@
-import style from "./Card.module.css";
-import img from "../../assets/images/web.png";
-import CardInfo from "../CardInfo/CardInfo";
-
-const Card = ({ type }) => {
-  return (
-    <div className={style.card}>
-      <div className={style.imageBox}>
-        <CardInfo type={type} />
-        <img className={style.img} width="400" height="272" src={img} alt="" />
-      </div>
-      <div className={style.description}>
-        <time className={style.date} dateTime="<2 августа, 2019>">
-          2 августа, 2019
-        </time>
-        <h2 className={style.title}>Национальное достояние – парки</h2>
-        <p className={style.text}>
-          В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться
-          система национальных парков – охраняемых территорий, где и сегодня каждый может
-          приобщиться к природе.
-        </p>
-        <a className={style.link} href="http://" target="_blank" rel="noopener noreferrer">
-          Лента.ру
-        </a>
-      </div>
-    </div>
-  );
-};
-
-export default Card;
