--- src/components/Loader/Loader.jsx
@@ -1,20 +1,0 @@
-import style from "./Loader.module.css";
-import { LoaderSvg } from "./loader.svg";
-
-const loaderSizes = {
-  small: 16,
-  medium: 24,
-  large: 80,
-};
-
-export const Loader = ({ size, inverse = false }) => {
-  const loaderColor = inverse ? "#fff" : "#888";
-  const stroke = size === "large" ? "1" : "2";
-
-  const wrapperStyleKey = "wrapper_" + size;
-  return (
-    <div className={style[wrapperStyleKey]}>
-      <LoaderSvg color={loaderColor} size={loaderSizes[size]} stroke={stroke} />
-    </div>
-  );
-};
