--- src/components/Registration/Registration.module.css
@@ -1,26 +1,0 @@
-.form {
-  width: 360px;
-  max-width: 100%;
-}
-.legend {
-  font: var(--font-Roboto-24-black);
-  margin-bottom: 22px;
-}
-.button {
-  margin-top: 40px;
-  width: 100%;
-}
-.footer {
-  display: flex;
-  gap: 8px;
-  justify-content: center;
-  flex-wrap: wrap;
-  margin-top: 16px;
-}
-.text {
-  font: var(--font-Inter-12-regular);
-}
-.changeFormButton {
-  font: var(--font-Inter-12-regular);
-  color: var(--color-main-blue);
-}
