+++ src/entities/CardInfo/index.ts
@@ -1,0 +1,1 @@
+export { default as CardInfo } from "./ui";
