+++ src/entities/CardInfo/ui/style.module.css
