+++ src/entities/input/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from "./model";
+export * from "../hooks";
