+++ src/entities/modal/ui/index.tsx
@@ -1,8 +1,8 @@
 import { FC, useEffect, useMemo } from "react";
 import ReactDOM from "react-dom";
-import { Overaly } from "../../../shared/ui";
+import Overaly from "shared/ui/Overlay/ui";
 import style from "./style.module.css";
-import icon from "../../../shared/assets/icons/close.svg";
+import icon from "shared/assets/icons/close.svg";
 import { IModalProps } from "../types";
 
 const Modal: FC<IModalProps> = ({ onClose, children }) => {
