+++ src/features/Card/index.ts
@@ -1,0 +1,1 @@
+export { default as Card } from "./ui";
