--- src/pages/Profile/Profile.jsx
@@ -1,13 +1,0 @@
-import CardList from "../../components/CardList/CardList";
-import ProfileInfo from "../../components/ProfileInfo/ProfileInfo";
-
-const Profile = () => {
-  return (
-    <>
-      <ProfileInfo />
-      <CardList type="profile" />
-    </>
-  );
-};
-
-export default Profile;
