--- src/pages/main/main.jsx
@@ -1,15 +1,0 @@
-import About from "../../components/About/About";
-import CardList from "../../components/CardList/CardList";
-import Intro from "../../components/Intro/Intro";
-
-const MainPage = () => {
-  return (
-    <>
-      <Intro />
-      <CardList />
-      <About />
-    </>
-  );
-};
-
-export default MainPage;
