+++ src/shared/ui/Loader/ui/loader.svg.js
