+++ src/shared/ui/Overlay/index.ts
@@ -1,0 +1,1 @@
+export { default as Overlay } from "./ui";
