+++ src/widgets/About/index.ts
@@ -1,0 +1,1 @@
+export { default as About } from "./ui";
