+++ src/widgets/CardList/index.ts
@@ -1,0 +1,1 @@
+export { default as CardList } from "./ui";
