+++ src/widgets/CardList/ui/style.module.css
