+++ src/widgets/Intro/index.ts
@@ -1,0 +1,1 @@
+export { default as Intro } from "./ui";
