+++ src/widgets/Intro/ui/style.module.css
