+++ src/widgets/Loading/index.ts
@@ -1,0 +1,1 @@
+export { default as Loading } from "./ui";
