+++ src/widgets/Loading/ui/style.module.css
