+++ src/widgets/NotFoundCards/index.ts
@@ -1,0 +1,1 @@
+export { default as NotFoundCards } from "./ui";
