+++ src/widgets/ProfileInfo/index.ts
@@ -1,0 +1,1 @@
+export { default as ProfileInfo } from "./ui";
