+++ src/widgets/ProfileInfo/ui/style.module.css
